refactor: 네비게이션 model 패키지 분리
