+++ android/app/src/main/java/com/yagubogu/ui/main/MainViewModel.kt
@@ -1,7 +1,7 @@
 package com.yagubogu.ui.main
 
 import androidx.lifecycle.ViewModel
-import com.yagubogu.ui.navigation.BottomNavKey
+import com.yagubogu.ui.navigation.model.BottomNavKey
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
