+++ android/app/src/main/java/com/yagubogu/ui/main/YaguBoguActivity.kt
@@ -29,7 +29,7 @@ import com.yagubogu.ui.login.auth.GoogleCredentialManager
 import com.yagubogu.ui.login.model.InAppUpdateType
 import com.yagubogu.ui.login.model.VersionInfo
 import com.yagubogu.ui.main.model.AutoLoginState
-import com.yagubogu.ui.navigation.Route
+import com.yagubogu.ui.navigation.model.Route
 import com.yagubogu.ui.navigation.YaguBoguRoute
 import com.yagubogu.ui.theme.YaguBoguTheme
 import dagger.hilt.android.AndroidEntryPoint
