+++ android/app/src/main/java/com/yagubogu/ui/main/component/MainNavigationBar.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.yagubogu.ui.navigation.BottomNavKey
+import com.yagubogu.ui.navigation.model.BottomNavKey
 import com.yagubogu.ui.theme.Gray200
 import com.yagubogu.ui.theme.Gray500
 import com.yagubogu.ui.theme.PretendardSemiBold12
