+++ android/app/src/main/java/com/yagubogu/ui/navigation/NavigationRoot.kt
@@ -12,12 +12,16 @@ import com.yagubogu.ui.favorite.FavoriteTeamScreen
 import com.yagubogu.ui.login.LoginScreen
 import com.yagubogu.ui.login.auth.GoogleCredentialManager
 import com.yagubogu.ui.main.MainScreen
+import com.yagubogu.ui.navigation.model.BottomNavKey
+import com.yagubogu.ui.navigation.model.Navigator
+import com.yagubogu.ui.navigation.model.Route
+import com.yagubogu.ui.navigation.model.toEntries
 import com.yagubogu.ui.setting.SettingScreen
 
 /**
  * 앱의 최상위 네비게이션 구조를 정의하는 루트 컴포저블.
  *
- * 각 경로([Route])에 따른 화면 컴포저블을 매핑하여 화면 전환을 관리합니다.
+ * 각 경로([com.yagubogu.ui.navigation.model.Route])에 따른 화면 컴포저블을 매핑하여 화면 전환을 관리합니다.
  *
  * @param googleCredentialManager 구글 인증 관리자
  * @param rootNavigator 최상위 라우팅 관리 Navigator
