+++ android/app/src/main/java/com/yagubogu/ui/navigation/model/BottomNavKey.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.ui.navigation
+package com.yagubogu.ui.navigation.model
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
